Add files via upload
Initial version of the program that takes json data (Not according to the protocol yet) and outputs:
time, velocity, altitude, acceleration, throttle %, propelent mass in S1 and S2, the velocity components, the velocity angle (compared to
the local horizontal), downrange distance, gravitational acceleration, centripital acceleration (or centrifugal? The one that points outwards), drag acceleration, potential energy (Compared to the ground) and vertical and horizontal kinetic energy.